Update regular expression for checksums to match package structure
Without this fix of the regex, the `validate-checksums` function will only
update the resources in the `boot-cljsjs-checksums.edn` file which are
matched by the incomplete regex. This leads to deletion of entries in
`boot-cljsjs-checksums.edn`.

The regex that describes which files have checksums now reflects:

- The cljsjs package structure uses `development` as directory name for
  dev resources ([documented here][1]).
- Resources that are checked do not only have an `.inc.js` file ending,
  the `.ext.js` resources are checked, too ([see here][2] or [here][3],
  for example).

[1]: https://github.com/cljsjs/packages/wiki/Creating-Packages#6-writing-the-package-task
[2]: https://github.com/cljsjs/packages/blob/4358108af31090b1dd5d4d9f3498689492771657/create-react-class/boot-cljsjs-checksums.edn
[3]: https://github.com/cljsjs/packages/blob/4358108af31090b1dd5d4d9f3498689492771657/parinfer-codemirror/boot-cljsjs-checksums.edn

Also, the regex that is displayed in the help message of
`validate-checksums` is now indentical to the one actually used.